Move fetch call for areas into neighborhood page component
